requirements.txt
apipkg==1.4
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
coverage==4.3.4
Django==1.11
django-rest-swagger==2.1.2
djangorestframework==3.6.2
execnet==1.4.1
factory-boy==2.8.1
Faker==0.7.11
httpie==0.9.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
openapi-codec==1.3.1
packaging==16.8
py==1.4.33
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==3.1.2
pytest-factoryboy==1.3.1
pytest-pythonpath==0.7.1
pytest-sugar==0.8.0
pytest-xdist==1.16.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
termcolor==1.1.0
uritemplate==3.0.0
